feat: Implement comments-counter.js
